requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-extensions==4.1
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
